Add files via upload
This was scrapped from the project. It was supposed to add another database of a person requesting a present from Santa, regardless if they were going to be on the nice/naughty list. Then, some other code (not created) was going to shift through the requester list and have a user input who was going to be on the nice/naughty list.